위치 표시 수정
