+++ app/src/main/java/campus/tech/kakao/map/SearchAdapter.kt
@@ -27,4 +27,5 @@ class SearchAdapter(
         notifyDataSetChanged()
     }
 
+
 }
