+++ app/src/main/res/layout/activity_place_item.xml
@@ -33,7 +33,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
-        android:layout_marginLeft="370dp"/>
+        androidLtextSize="10dp"
+        android:layout_marginLeft="350dp"/>
 
 
 </FrameLayout>
